+++ ini/features.ini
@@ -311,8 +311,9 @@ HAS_DUPLICATION_MODE                   = build_src_filter=+<src/gcode/control/M6
 SPI_FLASH_BACKUP                       = build_src_filter=+<src/gcode/control/M993_M994.cpp>
 PLATFORM_M997_SUPPORT                  = build_src_filter=+<src/gcode/control/M997.cpp>
 HAS_TOOLCHANGE                         = build_src_filter=+<src/gcode/control/T.cpp>
-FT_MOTION                              = build_src_filter=+<src/module/ft_motion.cpp> +<src/module/ft_motion> -<src/module/ft_motion/smoothing.cpp> +<src/gcode/feature/ft_motion>
+FT_MOTION                              = build_src_filter=+<src/module/ft_motion.cpp> +<src/module/ft_motion> -<src/module/ft_motion/smoothing.cpp> +<src/gcode/feature/ft_motion> -<src/gcode/feature/ft_motion/M495_M496.cpp>
 FTM_SMOOTHING                          = build_src_filter=+<src/module/ft_motion/smoothing.cpp>
+FTM_RESONANCE_TEST                     = build_src_filter=+<src/gcode/feature/ft_motion/M495_M496.cpp>
 HAS_LIN_ADVANCE_K                      = build_src_filter=+<src/gcode/feature/advance>
 PHOTO_GCODE                            = build_src_filter=+<src/gcode/feature/camera>
 CONTROLLER_FAN_EDITABLE                = build_src_filter=+<src/gcode/feature/controllerfan>
